refactor(higgs-tts): consolidate delay_pattern + bootstrap into utils.py
Two single-purpose modules with one function each, both shared across
the same set of stages. Fold into a single utils.py: the delay-pattern
helpers (apply / reverse + BOC_ID / EOC_ID) and truncate_rope_to_bf16.
